Move internal extensions to -*Platform specific files
