+++ library/controller/kmp-tor-controller/src/androidMain/kotlin/io/matthewnelson/kmp/tor/controller/TorController.kt
@@ -28,11 +28,11 @@ import io.matthewnelson.kmp.tor.controller.common.exceptions.TorControllerExcept
 import io.matthewnelson.kmp.tor.controller.common.file.Path
 import io.matthewnelson.kmp.tor.controller.common.internal.isUnixPath
 import io.matthewnelson.kmp.tor.controller.internal.controller.RealTorController
-import io.matthewnelson.kmp.tor.controller.internal.controller.getTorControllerDispatcher
+import io.matthewnelson.kmp.tor.controller.internal.getTorControllerDispatcher
 import io.matthewnelson.kmp.tor.controller.internal.io.ReaderWrapper
 import io.matthewnelson.kmp.tor.controller.internal.io.SocketWrapper
 import io.matthewnelson.kmp.tor.controller.internal.io.WriterWrapper
-import io.matthewnelson.kmp.tor.controller.internal.util.toTorController
+import io.matthewnelson.kmp.tor.controller.internal.toTorController
 import kotlinx.coroutines.withContext
 import java.io.*
 import java.net.Socket
