--- library/controller/kmp-tor-controller/src/commonMain/kotlin/io/matthewnelson/kmp/tor/controller/internal/controller/ControllerDispatchers.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) 2022 Matthew Nelson
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-package io.matthewnelson.kmp.tor.controller.internal.controller
-
-import kotlinx.coroutines.CloseableCoroutineDispatcher
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlin.jvm.JvmSynthetic
-
-@JvmSynthetic
-@Suppress("nothing_to_inline")
-@OptIn(ExperimentalCoroutinesApi::class)
-internal expect inline fun getTorControllerDispatcher(): CloseableCoroutineDispatcher
