+++ library/controller/kmp-tor-controller/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/controller/TorController.kt
@@ -27,8 +27,8 @@ import io.matthewnelson.kmp.tor.controller.common.file.Path
 import io.matthewnelson.kmp.tor.controller.common.internal.PlatformUtil
 import io.matthewnelson.kmp.tor.controller.common.internal.isUnixPath
 import io.matthewnelson.kmp.tor.controller.internal.controller.RealTorController
-import io.matthewnelson.kmp.tor.controller.internal.controller.getTorControllerDispatcher
-import io.matthewnelson.kmp.tor.controller.internal.util.toTorController
+import io.matthewnelson.kmp.tor.controller.internal.getTorControllerDispatcher
+import io.matthewnelson.kmp.tor.controller.internal.toTorController
 import kotlinx.coroutines.withContext
 import java.io.*
 import java.net.Socket
